+++ site/lupo.php
@@ -18,7 +18,7 @@ jimport('joomla.application.component.controller');
 $controller = JControllerLegacy::getInstance('Lupo');
 
 // Perform the Request task
-$controller->execute(JRequest::getCmd('task'));
+$controller->execute(JFactory::getApplication()->input->get('task'));
 
 // Redirect if set by the controller
 $controller->redirect();