+++ site/views/category/view.html.php
@@ -22,13 +22,32 @@ class LupoViewCategory extends JViewLegacy
 	// Overwriting JView display method
 	function display($tpl = null) 
 	{
+		$app        = JFactory::getApplication();
+
 		// Check for errors.
 		if (count($errors = $this->get('Errors'))) 
 		{
-			JError::raiseError(500, implode('<br />', $errors));
+			JFactory::getApplication()->enqueueMessage(implode('<br />', $errors), 'error');
 			return false;
 		}
-		
+
+		// Check for empty title and add site name if param is set
+		$title = $this->category['title'];
+		if (empty($title))
+		{
+			$title = $app->get('sitename');
+		}
+		elseif ($app->get('sitename_pagetitles', 0) == 1)
+		{
+			$title = JText::sprintf('JPAGETITLE', $app->get('sitename'), $title);
+		}
+		elseif ($app->get('sitename_pagetitles', 0) == 2)
+		{
+			$title = JText::sprintf('JPAGETITLE', $title, $app->get('sitename'));
+		}
+
+		JFactory::getDocument()->setTitle($title);
+
 		// Display the view
 		parent::display($tpl);
 	}